refactor(enrichment): extract transport, budget, and rendering into pr_reviewer modules
Closes #359
