+++ tests/test_context_budget.sh
@@ -82,26 +82,27 @@ check "output is valid UTF-8" \
 
 echo ""
 echo "=== Test: enrichment context trims are wired into the Python pipeline ==="
-# Enrichment rendering moved from scripts/sections/enrichment.sh into
-# scripts/run_enrichment.py (#7892). The skip-list and corpus-note behavior
-# must survive the migration.
-ENRICHMENT_PY="$ROOT_DIR/scripts/run_enrichment.py"
+# Enrichment rendering moved from scripts/sections/enrichment.sh into the
+# Python pipeline (#7892), and the render + skip-list logic was extracted from
+# scripts/run_enrichment.py into pr_reviewer/linked_sources.py (#359). The
+# skip-list and corpus-note behavior must survive both migrations.
+LINKED_SOURCES_PY="$ROOT_DIR/pr_reviewer/linked_sources.py"
 check "github.com raw HTML fetch is skipped" \
-  "$(grep -c 'Raw HTML fetch skipped for github.com' "$ENRICHMENT_PY")" "1"
+  "$(grep -c 'Raw HTML fetch skipped for github.com' "$LINKED_SOURCES_PY")" "1"
 check "non-github sources go through strip_source_to_text" \
-  "$(grep -c 'strip_source_to_text' "$ENRICHMENT_PY")" "2"
+  "$(grep -c 'strip_source_to_text' "$LINKED_SOURCES_PY")" "2"
 check "github.com is excluded from the parallel prefetch" \
-  "$(grep -c 'host == "github.com"' "$ENRICHMENT_PY")" "2"
+  "$(grep -c 'host == "github.com"' "$LINKED_SOURCES_PY")" "2"
 check "SKIP_FETCH_HOSTS constant is defined" \
-  "$(grep -c '^SKIP_FETCH_HOSTS' "$ENRICHMENT_PY")" "1"
+  "$(grep -c '^SKIP_FETCH_HOSTS' "$LINKED_SOURCES_PY")" "1"
 check "SKIP_FETCH_HOSTS list covers gitlab.com" \
-  "$(grep -c 'gitlab.com' "$ENRICHMENT_PY")" "1"
+  "$(grep -c 'gitlab.com' "$LINKED_SOURCES_PY")" "1"
 check "SKIP_FETCH_HOSTS list covers bitbucket.org" \
-  "$(grep -c 'bitbucket.org' "$ENRICHMENT_PY")" "1"
+  "$(grep -c 'bitbucket.org' "$LINKED_SOURCES_PY")" "1"
 check "Phase-1 loop skips SKIP_FETCH_HOSTS" \
-  "$(grep -c 'SKIP_FETCH_HOSTS' "$ENRICHMENT_PY")" "3"
+  "$(grep -c 'SKIP_FETCH_HOSTS' "$LINKED_SOURCES_PY")" "3"
 check "Phase-2 emits 'known non-Forgejo host' corpus note" \
-  "$(grep -c 'Raw HTML fetch skipped for known non-Forgejo host' "$ENRICHMENT_PY")" "1"
+  "$(grep -c 'Raw HTML fetch skipped for known non-Forgejo host' "$LINKED_SOURCES_PY")" "1"
 
 echo ""
 echo "=== Test: no tokens on curl argv in the incremental fetch ==="
@@ -109,8 +110,9 @@ check "incremental fetch passes the token via --config, not argv" \
   "$(grep -c 'Authorization: token \$GH_TOKEN" \\' "$ROOT_DIR/scripts/sections/config.sh" || true)" "0"
 check "incremental fetch uses curl_config_escape helper" \
   "$(grep -c 'curl_config_escape "Authorization: token' "$ROOT_DIR/scripts/sections/config.sh")" "1"
-# HTML stripping lives only in run_enrichment.py (via strip_source_text.py);
-# context.sh must not grow a parallel shell implementation again.
+# HTML stripping lives only in pr_reviewer/linked_sources.py (via
+# strip_source_text.py); context.sh must not grow a parallel shell
+# implementation again.
 check "context.sh has no parallel strip implementation" \
   "$(grep -c 'strip_source_to_text' "$ROOT_DIR/scripts/sections/context.sh" || true)" "0"
